Controladores agregados, tipos de datos arreglados y usando la opción AsSplitQuery

+++ Entities/Cargo.cs
@@ -5,7 +5,7 @@ namespace WebApiKalum_Backend.Entities
         public string CargoId { get; set; }
         public string Descripcion { get; set; }
         public string Prefijo { get; set; }
-        public double Monto { get; set; }
+        public Decimal Monto { get; set; }
         public bool GenerarMora { get; set; }
         public int PorcentajeMora { get; set; }
         public virtual List<CuentaPorCobrar> CuentasPorCobrar { get; set; }
