+++ Entities/InscripcionPago.cs
@@ -4,7 +4,7 @@ namespace WebApiKalum_Backend.Entities
     {
         public String BoletaPago { get; set; }
         public DateTime FechaPago { get; set; }
-        public Decimal Monto { get; set; }
+        public Double Monto { get; set; }
         public String Anio { get; set; }
         public String NoExpediente { get; set; }
         public virtual Aspirante Aspirante { get; set; }
